funclib.sh: use to SED for escaping in OpenSolaris shell
Reproducible in heirloom-sh on GNU/Linux, too.  There are two problems,
first that IFS='\' doesn't split the string properly:

    $ var='a\b'
    $ old_IFS=$IFS
    $ IFS='\'
    $ set dummy $var
    $ printf "%s\n" "$2"
    a\b

Second (and this is not happening in the OpenSolaris shell, just
heirloom-sh), IFS doesn't separate the fields properly when IFS is set
to non-whitespace character:

    $ var='a::b'  # three fields
    $ IFS=':'
    $ set dummy $var
    $ printf "%s\n" "$2" "$3"  # the empty field skipped
    a
    b

These problems declassify such shell to the slow $SED quote variant.

Originally reported in GNU Libtool bug #47000:
https://lists.gnu.org/archive/html/bug-libtool/2021-03/msg00000.html

* build-aux/funclib.sh (require_check_ifs_backslash): New helper
function to detect if '\' can be used as IFS field separator.
(func_quote_portable): Use $require_check_ifs_backslash and fallback to
SED escaping if '\' can not be used as IFS separator.
* bootstrap: Regenerate.